types changed at ProductBuilder

+++ src/main/java/com/orgabor/sportsgearstore/products/ProductBuilder.java
@@ -1,15 +1,16 @@
 package com.orgabor.sportsgearstore.products;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 public class ProductBuilder {
 
 	public static Product buildProduct(Map<String, Object> productParams) {
-		int productId = (int) productParams.get("productId");
+		long productId = (long) productParams.get("productId");
 		String name = (String) productParams.get("name");
 		String description = (String) productParams.get("description");
 		Categories category = (Categories) productParams.get("category");
-		double price = (double) productParams.get("price");
+		BigDecimal price = (BigDecimal) productParams.get("price");
 		int stock = (int) productParams.get("inStock");
 		String img = (String) productParams.get("img");
